+++ src/main/java/dev/flrp/economobs/hook/entity/ItemsAdderEntityHook.java
@@ -1,8 +1,8 @@
-package dev.flrp.economobs.hooks.entity;
+package dev.flrp.economobs.hook.entity;
 
 import dev.flrp.economobs.Economobs;
 import dev.flrp.economobs.configuration.Builder;
-import dev.flrp.economobs.listeners.ItemsAdderListener;
+import dev.flrp.economobs.listener.ItemsAdderListener;
 import dev.flrp.economobs.util.Methods;
 import dev.flrp.espresso.configuration.Configuration;
 import dev.flrp.espresso.hook.entity.custom.ItemsAdderEntityProvider;
