+++ src/main/java/dev/flrp/economobs/hook/stacker/StackMobListener.java
@@ -1,7 +1,7 @@
-package dev.flrp.economobs.hooks.stacker;
+package dev.flrp.economobs.hook.stacker;
 
 import dev.flrp.economobs.Economobs;
-import dev.flrp.economobs.hooks.SentinelHook;
+import dev.flrp.economobs.hook.SentinelHook;
 import dev.flrp.espresso.hook.entity.custom.EntityProvider;
 import dev.flrp.espresso.hook.stacker.StackMobStackerProvider;
 import org.bukkit.entity.LivingEntity;
@@ -33,7 +33,7 @@ public class StackMobListener extends StackMobStackerProvider {
         Player player = entity.getKiller();
         if(SentinelHook.isNPC(player)) player = plugin.getServer().getPlayer(SentinelHook.getNPCOwner(player));
         int stackSize = event.getDeathStep();
-        plugin.getRewardManager().handleLootReward(player, entity, plugin.getRewardManager().getLootContainer(entity.getType()), stackSize);
+        plugin.getRewardManager().handleLootReward(player, entity, plugin.getRewardManager().getLootContainer(entity.getType()), stackSize, entity.getType().name());
     }
 
 }
