+++ src/main/java/dev/flrp/economobs/hook/stacker/WildStackerListener.java
@@ -1,41 +1,40 @@
-package dev.flrp.economobs.hooks.stacker;
+package dev.flrp.economobs.hook.stacker;
 
+import com.bgsoftware.wildstacker.api.events.EntityUnstackEvent;
 import dev.flrp.economobs.Economobs;
-import dev.flrp.economobs.hooks.SentinelHook;
+import dev.flrp.economobs.hook.SentinelHook;
 import dev.flrp.espresso.hook.entity.custom.EntityProvider;
-import dev.flrp.espresso.hook.stacker.RoseStackerStackerProvider;
-import dev.rosewood.rosestacker.event.EntityUnstackEvent;
+import dev.flrp.espresso.hook.stacker.WildStackerStackerProvider;
 import org.bukkit.Bukkit;
+import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 
-public class RoseStackerListener extends RoseStackerStackerProvider {
+public class WildStackerListener extends WildStackerStackerProvider {
 
     private final Economobs plugin;
 
-    public RoseStackerListener(Economobs plugin) {
+    public WildStackerListener(Economobs plugin) {
         super(plugin);
         this.plugin = plugin;
     }
 
     @EventHandler
     public void onStackKill(EntityUnstackEvent event) {
-        LivingEntity entity = event.getStack().getEntity();
+        Entity source = event.getUnstackSource();
+        LivingEntity entity = event.getEntity().getLivingEntity();
 
         if(!plugin.getHookManager().getEntityProviders().isEmpty()) {
             for(EntityProvider provider : plugin.getHookManager().getEntityProviders()) if(provider.isCustomEntity(entity)) return;
         }
 
-        if(entity.getKiller() == null) return;
+        if(!(source instanceof Player)) return;
+        Player player = (Player) source;
         if(plugin.getConfig().getStringList("world-blacklist").contains(entity.getWorld().getName())) return;
         if(!plugin.getRewardManager().hasLootContainer(entity.getType())) return;
-
-        int before = event.getStack().getStackSize();
-        int after = event.getResult().getStackSize();
-        Player player = event.getStack().getEntity().getKiller();
         if(SentinelHook.isNPC(player)) player = Bukkit.getPlayer(SentinelHook.getNPCOwner(player));
-        plugin.getRewardManager().handleLootReward(player, entity, plugin.getRewardManager().getLootContainer(entity.getType()), before - after);
+        plugin.getRewardManager().handleLootReward(player, entity, plugin.getRewardManager().getLootContainer(entity.getType()), event.getAmount(), entity.getType().name());
     }
 
 }
