+++ src/main/java/dev/flrp/economobs/listener/ItemsAdderListener.java
@@ -1,8 +1,8 @@
-package dev.flrp.economobs.listeners;
+package dev.flrp.economobs.listener;
 
 import dev.flrp.economobs.Economobs;
-import dev.flrp.economobs.hooks.SentinelHook;
-import dev.flrp.economobs.hooks.entity.ItemsAdderEntityHook;
+import dev.flrp.economobs.hook.SentinelHook;
+import dev.flrp.economobs.hook.entity.ItemsAdderEntityHook;
 import dev.flrp.espresso.hook.entity.custom.EntityType;
 import dev.flrp.espresso.table.LootContainer;
 import dev.lone.itemsadder.api.Events.CustomEntityDeathEvent;
@@ -38,7 +38,7 @@ public class ItemsAdderListener implements Listener {
         if(SentinelHook.isNPC(player)) player = Bukkit.getPlayer(SentinelHook.getNPCOwner(player));
         LootContainer lootContainer = itemsAdderHook.hasLootContainer(entityName)
                 ? itemsAdderHook.getLootContainer(entityName) : itemsAdderHook.getDefaultLootContainer();
-        plugin.getRewardManager().handleCustomEntityLootReward(player, (LivingEntity) entity, lootContainer, entityName);
+        plugin.getRewardManager().handleLootReward(player, (LivingEntity) entity, lootContainer, 1, entityName);
     }
 
 }
