+++ src/main/java/dev/flrp/economobs/manager/HookManager.java
@@ -4,8 +4,8 @@ import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import dev.flrp.economobs.Economobs;
 import dev.flrp.economobs.configuration.Locale;
-import dev.flrp.economobs.hooks.entity.InfernalMobsHook;
-import dev.flrp.economobs.hooks.entity.LevelledMobsHook;
+import dev.flrp.economobs.hook.entity.InfernalMobsHook;
+import dev.flrp.economobs.hook.entity.LevelledMobsHook;
 import dev.flrp.espresso.hook.economy.EconomyProvider;
 import dev.flrp.espresso.hook.economy.EconomyType;
 import dev.flrp.espresso.hook.entity.custom.EntityProvider;
