+++ src/main/java/dev/flrp/economobs/module/EntityModule.java
@@ -3,8 +3,8 @@ package dev.flrp.economobs.module;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
 import dev.flrp.economobs.Economobs;
-import dev.flrp.economobs.hooks.entity.ItemsAdderEntityHook;
-import dev.flrp.economobs.hooks.entity.MythicMobsEntityHook;
+import dev.flrp.economobs.hook.entity.ItemsAdderEntityHook;
+import dev.flrp.economobs.hook.entity.MythicMobsEntityHook;
 import dev.flrp.espresso.hook.entity.custom.EntityProvider;
 import dev.flrp.espresso.hook.entity.custom.ItemsAdderEntityProvider;
 import dev.flrp.espresso.hook.entity.custom.MythicMobsEntityProvider;
