+++ src/main/java/dev/flrp/economobs/module/ItemModule.java
@@ -3,9 +3,9 @@ package dev.flrp.economobs.module;
 import com.google.inject.AbstractModule;
 import com.google.inject.multibindings.Multibinder;
 import dev.flrp.economobs.Economobs;
-import dev.flrp.economobs.hooks.item.ItemsAdderItemHook;
-import dev.flrp.economobs.hooks.item.MMOItemsItemHook;
-import dev.flrp.economobs.hooks.item.OraxenItemHook;
+import dev.flrp.economobs.hook.item.ItemsAdderItemHook;
+import dev.flrp.economobs.hook.item.MMOItemsItemHook;
+import dev.flrp.economobs.hook.item.OraxenItemHook;
 import dev.flrp.espresso.hook.item.ItemProvider;
 import dev.flrp.espresso.hook.item.ItemsAdderItemProvider;
 import dev.flrp.espresso.hook.item.MMOItemsItemProvider;
